#239 [fix]: 시스템 알림 권한 로그인 후, 띄우는 것으로 변경
