+++ app/src/main/java/com/byeboo/app/presentation/splash/SplashState.kt
@@ -7,5 +7,6 @@ sealed interface SplashStateSideEffect {
     data object NavigateToTermsOfService : SplashStateSideEffect
     data object StartKakaoTalkLogin : SplashStateSideEffect
     data object StartKakaoWebLogin : SplashStateSideEffect
+    data object RequestNotificationPermission : SplashStateSideEffect
     data class ShowSnackBar(val message: String) : SplashStateSideEffect
 }
